refactor enemy drops are in enemy class
