+++ app/classes/Enemy.js
@@ -1,7 +1,8 @@
 import entityPrefab from './entityPrefab'
-import { socket } from '../sockets'
+import { GameItems, socket } from '../sockets'
 import HealthBar from '../states/utils/HealthBar.js'
 import enemyProperties from '../properties/enemyProperties'
+import Loot from 'APP/app/classes/Loot'
 
 /* global StackQuest, Phaser */
 
@@ -107,6 +108,14 @@ export default class Enemy extends entityPrefab {
   }
 
   dropLoot() {
-
+    const chance = Math.floor(Math.random() * 100)
+    if (chance < 20) {
+      const newItemName = Math.random().toString(36).substr(2, 5) // need this in order to create a random item name
+      const itemTypes = ['weapon', 'armor', 'loot']
+      const itemType = itemTypes[Math.floor(Math.random() * itemTypes.length)]
+      GameItems[newItemName] = new Loot(StackQuest.game, newItemName, { x: this.x, y: this.y }, itemType)
+      const newItem = GameItems[newItemName]
+      socket.emit('createItem', { itemPos: newItem.position, name: newItem.name, key: newItem.key })
+    }
   }
 }
